# docs/swagger.yaml
definitions:
...
  request.IdsReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
  response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  system.SysUserLogin:
    properties:
      agent:
        description: 代理
        type: string
      createdAt:
        description: 创建时间
        type: string
      error_message:
        description: 错误信息
        type: string
      id:
        description: 主键ID
        type: integer
      ip:
        description: 请求ip
        type: string
      status:
        description: 登录状态
        type: integer
      updatedAt:
        description: 更新时间
        type: string
      userName:
        description: 用户登录名
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample Server pets
  title: Swagger Example API
  version: 0.0.1
paths:
...
  /sysUserLogin/createSysUserLogin:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建SysUserLogin
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.SysUserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: 创建SysUserLogin
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建SysUserLogin
      tags:
      - SysUserLogin
  /sysUserLogin/deleteSysUserLogin:
    delete:
      consumes:
      - application/json
      parameters:
      - description: SysUserLogin模型
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.SysUserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: 删除SysUserLogin
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除SysUserLogin
      tags:
      - SysUserLogin
  /sysUserLogin/deleteSysUserLoginByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 批量删除SysUserLogin
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: 批量删除SysUserLogin
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 批量删除SysUserLogin
      tags:
      - SysUserLogin
  /sysUserLogin/findSysUserLogin:
    get:
      consumes:
      - application/json
      parameters:
      - description: 代理
        in: query
        name: agent
        type: string
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - description: 错误信息
        in: query
        name: error_message
        type: string
      - description: 主键ID
        in: query
        name: id
        type: integer
      - description: 请求ip
        in: query
        name: ip
        type: string
      - description: 登录状态
        in: query
        name: status
        type: integer
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      - description: 用户登录名
        in: query
        name: userName
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 用id查询SysUserLogin
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 用id查询SysUserLogin
      tags:
      - SysUserLogin
  /sysUserLogin/getSysUserLoginList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 代理
        in: query
        name: agent
        type: string
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - description: 错误信息
        in: query
        name: error_message
        type: string
      - description: 主键ID
        in: query
        name: id
        type: integer
      - description: 请求ip
        in: query
        name: ip
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - description: 登录状态
        in: query
        name: status
        type: integer
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      - description: 用户登录名
        in: query
        name: userName
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 分页获取SysUserLogin列表,返回包括列表,总数,页码,每页数量
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取SysUserLogin列表
      tags:
      - SysUserLogin
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"